Added WhenExisting
It has to be checked whether this could be a useful alternative to
calling update() and overwrite().
One (small) drawback is that it is not possible to throw an exception
when the user wants an update, but the file to update does not exist.
This could be handled by adding a value like WhenMissing.ONLY_UPDATE,
but this gets complicated and less readable compared to just calling
update()